+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All notable changes to the "Hydro Helper" extension will be documented in this file.
 
-## v0.1.2
+## v0.1.4
 
-- **Add webviewColumn setting to configure webview display position**
-- Refactor score and status checks to use strict undefined comparisons
+- Fix solution reply rendering
+- Avoid cache video files
+- Refactor code
 
 Want to see the previous log? [Click here](https://github.com/langningchen/hydro-helper/commits/main/CHANGELOG.md)